eventsRepository Adds the following handlers: - Clicking the action button - Window click to clear the action menu - Handlers for each menu option in the action menu
elementCreator
- Exposes function to create an Action Menu

tableElementManipulation
- Appends a whole action menu to the last edge cell

tableObjectManipulation
-Generates a random id for rows
- Defines functions to covert a row array to JSON and CSV format

styles:
- Sets styles for the displayable action menu.

+++ modules/eventsRepository.js
@@ -1,20 +1,26 @@
 import { tableSearch, toggleColumnBtn, toggleDeleteBtn } from './tableElementManipulation.js';
-import { addRow, addColumn, updateCell, getTableObject, setTableObject, deleteRecord } from './tableObjectManipulation.js'
+import { addRow, addColumn, updateCell, getTableObject, setTableObject, deleteRecord, getRowJson, getRowCsv } from './tableObjectManipulation.js'
 import { getSortOrder, toggleSortOrder } from './utilityScript.js';
 
 function setGlobalEventListener(type, selector, func) {
   document.addEventListener(type, e => {
+    // console.log(e.target, selector, e.target.matches(selector), func);
     if (e.target.matches(selector)) func(e);
   });
 }
 
-const makeEditable = e => {
-  e.target.contentEditable = 'true';
+const makeElementEditable = element => {
+  element.contentEditable = true;
+}
+
+const onDataCellDoubleClick = e => {
+  // e.target.contentEditable = 'true';
+  makeElementEditable(e.target);
 
   e.target.focus({focusVisible: true});
 };
 
-const updateCellText = e => {
+const onDataCellKeyPress = e => {
   
   if (e.key === 'Enter') {
     e.preventDefault()
@@ -22,14 +28,15 @@ const updateCellText = e => {
     let elementText = e.target.textContent;
     const cellType = e.target.getAttribute('type');
     let cellColor = 'black';
-    let elementId = e.target.className;
+    // let elementId = e.target.className;
+    let elementId = e.target.closest('[class]').classList[0];
     const fatherId = e.target.closest('tr').id;
 
     e.target.contentEditable = 'false';
 
     const errorHandler = (errorMsg) => {
       alert(errorMsg);
-      makeEditable(e);
+      onDataCellDoubleClick(e);
       e.target.style.borderColor = 'red';
     }
 
@@ -51,6 +58,7 @@ const updateCellText = e => {
           errorHandler('Not a valid Email');
           return;
         }
+        console.log('i am here');
         // elementId = e.target.parentElement.getAttribute('id');
         break;
       case 'url' :
@@ -61,6 +69,7 @@ const updateCellText = e => {
         // elementId = e.target.parentElement.getAttribute('id');
         break;
     } 
+    // console.log(elementText, fatherId, elementId);
     updateCell(fatherId, elementId, elementText);
     e.target.style.borderColor = cellColor;
   }
@@ -115,7 +124,7 @@ const onHeaderClick = e => {
 
 const dataCheckboxSelector = 'input[name="row-checker"]';
 
-const checkBoxesQuery = () => {
+const queryCheckBoxes = () => {
   const allCheckboxes = document.querySelectorAll(dataCheckboxSelector);
   let areAllUnchecked = true;
 
@@ -157,7 +166,7 @@ const onDataRowsCheckboxClick = e => {
   }
 
   elementRow.classList.remove('highlighted-row');
-  if (checkBoxesQuery()) toggleDeleteBtn(isElementChecked);
+  if (queryCheckBoxes()) toggleDeleteBtn(isElementChecked);
 };
 
 function setCellsEventListeners() {
@@ -167,8 +176,8 @@ function setCellsEventListeners() {
   const headerCheckboxSelector = 'input[name="all-rows-checker"]';
   // console.log(document.querySelectorAll(headerCheckboxSelector), document.querySelectorAll(dataCheckboxSelector));
 
-  setGlobalEventListener('dblclick', editableElementsSelector, makeEditable);
-  setGlobalEventListener('keydown', editableElementsSelector, updateCellText);
+  setGlobalEventListener('dblclick', editableElementsSelector, onDataCellDoubleClick);
+  setGlobalEventListener('keydown', editableElementsSelector, onDataCellKeyPress);
   setGlobalEventListener('click', anchorSelector, onAnchorClick);
   setGlobalEventListener('click', headerSelector, onHeaderClick);
   setGlobalEventListener('click', headerCheckboxSelector, onFirstCheckboxClick);
@@ -206,12 +215,29 @@ const onModalCloseClick = (e) => {
 const onDeleteBtnClick = e => {
   const allCheckboxes = document.querySelectorAll(dataCheckboxSelector);
   
-  allCheckboxes.forEach (checkbox => {
-    if (checkbox.checked) {
-      const rowId = checkbox.closest('tr').id;
-      deleteRecord(rowId);
+  if(confirm('\nThis action will delete the selected row\\s.\n\nDo you wish to continue?')){
+    allCheckboxes.forEach (checkbox => {
+      if (checkbox.checked) {
+        const rowId = checkbox.closest('tr').id;
+        deleteRecord(rowId);
+      }
+    });
+    toggleDeleteBtn(false);
+  }
+};
+
+const onActionBtnClick = e => {
+  // const rowMenu = document.getElementsByClassName('menu-options');
+  const optionsMenus = document.getElementsByClassName('menu-options');
+  for (let i = 0; i < optionsMenus.length; i++) {
+    const openMenu = optionsMenus[i];
+    if (openMenu.classList.contains('open-menu')) {
+      openMenu.classList.remove('open-menu');
     }
-  });
+  }
+
+  const rowId = e.target.closest('tr').id;
+  document.getElementById(`m${rowId}`).classList.toggle('open-menu');
 };
 
 function setBtnsEventListeners() {
@@ -220,18 +246,83 @@ function setBtnsEventListeners() {
   setGlobalEventListener('click', '#modal-submit-btn', onModalSubmitClick);
   setGlobalEventListener('click', '#modal-close-btn', onModalCloseClick);
   setGlobalEventListener('click', '#delete-btn', onDeleteBtnClick);
+  setGlobalEventListener('click', '.action-btn', onActionBtnClick);
 }
 
 const onInput = (e) => {
   tableSearch(e.target.value, getTableObject());
 };
 
+const onWindowClick = e => {
+  if (!e.target.matches('.action-btn')) {
+    const optionsMenus = document.getElementsByClassName('menu-options');
+
+    for (let i = 0; i < optionsMenus.length; i++) {
+      var openMenu = optionsMenus[i];
+      if (openMenu.classList.contains('open-menu')) {
+        openMenu.classList.remove('open-menu');
+      }
+    }
+  }
+};
+
 function setOtherEventListeners() {
   setGlobalEventListener('input', '#search', onInput);
+  window.addEventListener('click', onWindowClick);
 }
+
+const onEditMenuClick = e => {
+  // console.log('im in')
+  const elementRowChildren = e.target.closest('tr').children;
+  // console.log(elementRowChildren);
+  
+  let i = 1
+  for (i = 1; i < elementRowChildren.length - 2; i++) {
+    makeElementEditable(elementRowChildren[i]);
+  }
+  makeElementEditable(elementRowChildren[i].children[0]);
+};
+
+const onDeleteMenuClick = e => {
+  if(confirm('\nThis action will delete the selected row.\n\nDo you wish to continue?')) {
+    const elementRowId = e.target.closest('tr').id;
+    deleteRecord(elementRowId);
+  }
+};
+
+const onCopyMenuClick = e => {
+  /* TODO implement document.execCommand for better compatibility*/
+  const elementRowId = e.target.closest('tr').id;
+  console.log('i am');
+  const rowJson = (e.target.className === 'copy-csv') ? getRowCsv(elementRowId) : getRowJson(elementRowId);
+  if (navigator.clipboard) {
+    console.log('copying');
+    navigator.clipboard.writeText(rowJson).then(
+      () => console.log('Row Copied'),
+      () => console.log('Row could not be copied')
+    );
+  }
+}
+
+function setRowMenuEventListeners() {
+  const editOptionSelector = 'a.edit';
+  const copyJsonOptionSelector = 'a.copy-json';
+  const copyCsvOptionSelector = 'a.copy-csv';
+  const deleteOptionSelector = 'a.delete';
+
+  // console.log('inRowmenu')
+  setGlobalEventListener('click', editOptionSelector, onEditMenuClick);
+  setGlobalEventListener('click', deleteOptionSelector, onDeleteMenuClick);
+  setGlobalEventListener('click', copyJsonOptionSelector, onCopyMenuClick);
+  setGlobalEventListener('click', copyCsvOptionSelector, onCopyMenuClick);
+}
+
+
+
 
 export function setAllEventListeners() {
   setCellsEventListeners();
   setBtnsEventListeners();
   setOtherEventListeners();
+  setRowMenuEventListeners();
 }
